requirements.txt
# local package
# -e .

# external requirements
click==7.1.2
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1


# backwards compatibility
pathlib2

# project requirements
pandas==1.1.1
requests==2.24.0
tqdm==4.48.2
texthero==1.0.9
lxml==4.5.2